backend/requirements.txt

  
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.10.3
requests==2.22.0
stripe
toml==0.9.6
urllib3==1.25.8
Werkzeug==1.0.1